requirements.txt
flask
pywebview
pyinstaller
typing-extensions
